Rename oracle service, fix fee RPCs to align with geth
