+++ lib/ain-cpp-imports/src/bridge.rs
@@ -37,7 +37,7 @@ pub mod ffi {
         fn getNetwork() -> String;
         fn getEthMaxConnections() -> u32;
         fn getEthMaxResponseByteSize() -> u32;
-        fn getSuggestedPriorityFeePercentile() -> u32;
+        fn getSuggestedPriorityFeePercentile() -> i64;
         fn getDifficulty(block_hash: [u8; 32]) -> u32;
         fn getChainWork(block_hash: [u8; 32]) -> [u8; 32];
         fn getPoolTransactions() -> Vec<TransactionData>;
