+++ lib/ain-grpc/src/rpc/eth.rs
@@ -228,7 +228,7 @@ pub trait MetachainRPC {
         &self,
         block_count: U256,
         newest_block: BlockNumber,
-        reward_percentile: Vec<usize>,
+        reward_percentile: Vec<i64>,
     ) -> RpcResult<RpcFeeHistory>;
 
     #[method(name = "maxPriorityFeePerGas")]
@@ -898,10 +898,9 @@ impl MetachainRPCServer for MetachainRPCModule {
     }
 
     fn gas_price(&self) -> RpcResult<U256> {
-        let curr_block = self.get_block(Some(BlockNumber::Latest))?.header.number;
         self.handler
-            .block
-            .get_legacy_fee(curr_block)
+            .oracle
+            .suggest_legacy_fee()
             .map_err(to_custom_err)
     }
 
@@ -933,14 +932,14 @@ impl MetachainRPCServer for MetachainRPCModule {
         &self,
         block_count: U256,
         newest_block: BlockNumber,
-        reward_percentile: Vec<usize>,
+        reward_percentile: Vec<i64>,
     ) -> RpcResult<RpcFeeHistory> {
         let highest_block_number = self.get_block(Some(newest_block))?.header.number;
         let attrs = ain_cpp_imports::get_attribute_values(None);
 
         let fee_history = self
             .handler
-            .block
+            .oracle
             .fee_history(
                 block_count,
                 highest_block_number,
@@ -953,10 +952,9 @@ impl MetachainRPCServer for MetachainRPCModule {
     }
 
     fn max_priority_fee_per_gas(&self) -> RpcResult<U256> {
-        let curr_block = self.get_block(Some(BlockNumber::Latest))?.header.number;
         self.handler
-            .block
-            .suggested_priority_fee(curr_block)
+            .oracle
+            .suggest_priority_fee()
             .map_err(to_custom_err)
     }
 
@@ -978,7 +976,7 @@ impl MetachainRPCServer for MetachainRPCModule {
                     .map(|block| block.header.number)
                     .ok_or(RPCError::BlockNotFound)?;
 
-                let starting_block = self.handler.block.get_starting_block_number();
+                let starting_block = self.handler.oracle.get_starting_block_number();
 
                 let highest_block = current_block + (highest_native_block - current_native_height); // safe since current height cannot be higher than seen height
                 debug!(target:"rpc", "Highest native: {highest_native_block}\nCurrent native: {current_native_height}\nCurrent ETH: {current_block}\nHighest ETH: {highest_block}");
