+++ lib/ain-grpc/src/subscription/eth.rs
@@ -134,7 +134,7 @@ impl MetachainPubSubServer for MetachainPubSubModule {
                                         .ok_or(format_err!("Unable to find latest block"))?
                                         .header
                                         .number;
-                                    let starting_block = handler.block.get_starting_block_number();
+                                    let starting_block = handler.oracle.get_starting_block_number();
                                     let highest_block = current_block + (highest - current); // safe since current can never be more than highest
 
                                     Ok(PubSubSyncStatus::Detailed(SyncStatusMetadata {
