+++ src/ffi/ffiexports.h
@@ -16,7 +16,7 @@ static constexpr uint32_t DEFAULT_ETH_MAX_CONNECTIONS = 100;
 static constexpr uint32_t DEFAULT_ETH_MAX_RESPONSE_SIZE_MB = 25;  // 25 megabytes
 
 // Defaults for attributes relating to gasprice oracle settings
-static constexpr uint32_t DEFAULT_SUGGESTED_PRIORITY_FEE_PERCENTILE = 60;
+static constexpr int64_t DEFAULT_SUGGESTED_PRIORITY_FEE_PERCENTILE = 60;
 
 static constexpr uint32_t DEFAULT_ECC_LRU_CACHE_COUNT = 10000;
 static constexpr uint32_t DEFAULT_EVMV_LRU_CACHE_COUNT = 10000;
@@ -73,7 +73,7 @@ rust::string getNetwork();
 uint32_t getDifficulty(std::array<uint8_t, 32> blockHash);
 uint32_t getEthMaxConnections();
 uint32_t getEthMaxResponseByteSize();
-uint32_t getSuggestedPriorityFeePercentile();
+int64_t getSuggestedPriorityFeePercentile();
 std::array<uint8_t, 32> getChainWork(std::array<uint8_t, 32> blockHash);
 rust::vec<TransactionData> getPoolTransactions();
 uint64_t getNativeTxSize(rust::Vec<uint8_t> rawTransaction);
